Select de disciplinas atualiza conforme o curso selecionado no CRUD de Questões

+++ src/main/java/br/edu/fatecguarulhos/projetoavalia/repository/DisciplinaRepository.java
@@ -8,5 +8,5 @@ import br.edu.fatecguarulhos.projetoavalia.model.entity.Disciplina;
 
 public interface DisciplinaRepository extends JpaRepository<Disciplina, Integer>{
 
-	List<Disciplina> findByCursoId(Long cursoId);
+	List<Disciplina> findByCursoId(int cursoId);
 }
